separate all admin view component to one blade file for each component
