+++ resources/views/admin/layouts/footer.blade.php
@@ -1,0 +1,9 @@
+<footer class="site-footer">
+    <div class="footer-inner bg-white">
+        <div class="row">
+            <div class="col-sm-6">
+                Copyright &copy; 2018 <a href="{{ url('/')}}">UlosZone</a>
+            </div>
+        </div>
+    </div>
+</footer>
