+++ resources/views/admin/roles/index.blade.php
@@ -1,4 +1,4 @@
-@extends('admin.layout.app')
+@extends('admin.layouts.app')
 
 @section('content')
 <div class="container">
